requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
supabase>=2.4.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pytest>=8.0.0
